stats/requirements.txt
flask
rcon
python-a2s